Damageformula (#814)
* added additional effects to npcs

This change requires rerunning GameDataParser
dummy wont die now

* skill break added

* multiple combat changes

added proximity tracking
fixed conditional stats
added a multiple skill cast state tracker
added dependOnCasterState support for early region skill canceling
started working on illusory blades support

* typo fix

* rb audit changes

* runeblade base kit feature complete

* clean up

* runeblade skill tweaking

runeblade is nearly feature complete, just need to implement nairin lapenshard

* implemented sensor skills, fixed conditional passive states, tweaked sp regen

* formatting

* clean up

* feedback adjustments

* feedback

* moving constants around

* compile fix

* format

* feedback

* changing skill level from int to short

+++ Maple2Storage/Types/Metadata/ItemMetadata.cs
@@ -132,7 +132,7 @@ public class ItemSkillMetadata
     [XmlElement(Order = 1)]
     public int SkillId;
     [XmlElement(Order = 2)]
-    public int SkillLevel;
+    public short SkillLevel;
 }
 
 [XmlType]
@@ -289,7 +289,7 @@ public class ItemAdditionalEffectMetadata
     [XmlElement(Order = 1)]
     public int[] Id;
     [XmlElement(Order = 2)]
-    public int[] Level;
+    public short[] Level;
 }
 
 [XmlType]
